v1.1
- Add function in function execution.
- calculate makes calculations in float type

+++ src/ru/jenya705/component/GlobalComponent.java
@@ -7,15 +7,16 @@ public class GlobalComponent {
 		String exercise = data.getArgs()[0];
 		String operators[] = exercise.split("[0-9]+");
 		String operands[] = exercise.split("[+-/*%]");
-		int result = Integer.parseInt(operands[0]);
+		
+		float result = Float.parseFloat(operands[0]);
 		for (int i = 1; i < operands.length; ++i) {
-			if (operators[i].equals("+")) result += Integer.parseInt(operands[i]);
-			else if (operators[i].equals("-")) result -= Integer.parseInt(operands[i]);
-			else if (operators[i].equals("*")) result *= Integer.parseInt(operands[i]);
-			else if (operators[i].equals("/")) result /= Integer.parseInt(operands[i]);
-			else result %= Integer.parseInt(operands[i]);
+			if (operators[i].equals("+")) result += Float.parseFloat(operands[i]);
+			else if (operators[i].equals("-")) result -= Float.parseFloat(operands[i]);
+			else if (operators[i].equals("*")) result *= Float.parseFloat(operands[i]);
+			else if (operators[i].equals("/")) result /= Float.parseFloat(operands[i]);
+			else result %= (int) Float.parseFloat(operands[i]);
 		}
-		data.setResult(Integer.toString(result));
+		data.setResult(Float.toString(result));
 		
 	}
